+++ history.rst
@@ -6,6 +6,19 @@ by Gerold Penz 2013
 
 
 =============
+Version 0.1.0
+=============
+
+2013-06-23
+
+- Error module created
+
+- Responses module created
+
+- Base structure
+
+
+=============
 Version 0.0.1
 =============